+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 	(at the beginning of a new line )
 -->
 
+## __WORK IN PROGRESS__
+* (AlCalzone) Add support for connectionType and dataSource (fixes #386)
+
 ## 1.20.1 (2020-01-25)
 * (AlCalzone) Add the `vehicle` adapter type (fixes #384)
 * (AlCalzone) The auto-publish on Github Actions now also works for pre-releases, e.g. `v1.0.0-beta.2` (#387)